sonounoweb/requirements.txt
asgiref==3.8.1
Django==5.0.2
Pillow==10.3.0
sqlparse==0.4.4
tzdata==2024.1
python-dotenv==1.0.1
matplotlib==3.8.3
pygame==2.5.2
scipy==1.12.0
pandas==2.2.1